Refactor Board.prototype.countChecks to return boolean
Number of checks no longer used by anything in the program.
It was done that way in case of behaviour specific to double checks, but
now double checks are handled the same as any move/check - before any
"valid" player move, the move is simulated and the "new" board checked
if the player king will be in check.
Double checks will have this remain true unless the king moves out of
check, so there's no need for the number of checks to be found.