Fix conversion of non-capture pawn moves across different files
e2->f4 would otherwise convert to f4 and push the f-pawn which is
unintentional. This should be prevented at the conversion level because
there is no disambiguation between files for non-capture pawn moves.

Invalid captures can still be converted as there is enough
disambiguation in the algebraic to fail when the move is validated in
the player move method.